backend/fast_api/requirements.txt
fastapi
uvicorn[standard]


openai~=1.55.3
PyYAML~=6.0.1

numpy~=2.1.1
torch~=2.4.1
transformers~=4.44.2
fugashi[unidic-lite]~=1.3.2

flake8>=5.0.0
flake8-docstrings>=1.7.0
